docs - Add project overview as body matter of README
Before this commit, the README was incomplete. Now, this file outlines
the main points of the project.

Watch out for big changes to this file. This commit represents
the 80% mark for the Minimum Viable Product (MVP). So, major changes,
especially to the list of features, indicate that it's time to
re-evaluate this project's relative priority to other projects.